restructured code & autostop based on os
