+++ src/utilities/getInlineCompletionItems.ts
@@ -1,67 +1,14 @@
-import * as vscode from "vscode";
 import CodeAssistantInlineCompletionItem from "../inlineSuggestions/inlineCompletionItem";
-import runCompletion from "./runCompletion";
-import { AutocompleteResult, ResultEntry } from "../binary/requests/requests";
-import { isMultiline } from "./utils";
+import { createDecorationType, extractCompletionParts, toCompletionItem } from "./utils";
 import handleGetCompletion from "../commands/handleGetCompletion";
+import runCompletion from "./runCompletion";
+import * as vscode from "vscode";
+import * as os from 'os';
 
 const INLINE_REQUEST_TIMEOUT = 3000;
 
 // this will be used to collect the chunks of streaming data
 const callsForCompletions = new Map<string, CodeAssistantInlineCompletionItem>();
-
-function createDecorationType(): vscode.TextEditorDecorationType {
-  return vscode.window.createTextEditorDecorationType({
-    after: {
-      margin: "0 0 0 1ch",
-      color: new vscode.ThemeColor("editorGhostText.foreground"),
-    },
-    rangeBehavior: vscode.DecorationRangeBehavior.ClosedClosed,
-  });
-}
-
-function extractCompletionParts(text: string): { before: string; after: string } {
-  const lastNewline = text.lastIndexOf("\n");
-  if (lastNewline === -1) return { before: "", after: text };
-  return {
-    before: text.slice(0, lastNewline + 1),
-    after: text.slice(lastNewline + 1),
-  };
-}
-
-function calculateRange(
-  position: vscode.Position,
-  response: AutocompleteResult,
-  result: ResultEntry
-): vscode.Range {
-  return new vscode.Range(
-    position.translate(0, -response.old_prefix.length),
-    isMultiline(result.old_suffix)
-      ? position
-      : position.translate(0, result.old_suffix.length)
-  );
-}
-
-function toCompletionItem(
-  insertText: string,
-  position: vscode.Position,
-  autoCompleteResult: AutocompleteResult,
-  resultEntry: ResultEntry
-) {
-  return new CodeAssistantInlineCompletionItem(
-    insertText,
-    resultEntry,
-    calculateRange(position, autoCompleteResult, resultEntry),
-    undefined,
-    resultEntry.completion_metadata?.model_id,
-    resultEntry.completion_metadata?.prompt_id,
-    resultEntry.completion_metadata?.input,
-    resultEntry.completion_metadata?.output,
-    resultEntry.completion_metadata?.completion_kind,
-    resultEntry.completion_metadata?.is_cached,
-    resultEntry.completion_metadata?.snippet_context
-  )
-}
 
 export default async function getInlineCompletionItems(
   document: vscode.TextDocument,
@@ -87,9 +34,10 @@ export default async function getInlineCompletionItems(
   let lastResolvedText = "";
   let cancelled = false;
 
-  const docChangeDisposable = vscode.workspace.onDidChangeTextDocument(e => {
-    if (e.document === document) cancelled = true;
-  });
+  const docChangeDisposable = (os.arch() === 'arm64')? 
+    vscode.workspace.onDidChangeTextDocument(e => cancelled=(e.document === document) )
+    :vscode.window.onDidChangeTextEditorSelection(e => cancelled=(e.textEditor.document === document) );
+  
 
   const isEmptyLine = document.lineAt(position.line).text.trim().length === 0;
   const completionId = `${document.uri.toString()}-${position.line}-${position.character}`;
@@ -110,7 +58,7 @@ export default async function getInlineCompletionItems(
 
     // loop through streaming data
     for await (let chunk of completionGenerator) {
-      if (cancelled) break;
+      if (cancelled) return;
       const result = chunk?.results[0]
       if (!result) return;
